Give type tags an order-independent option pack
A tag's size moves off its positional slot into a trailing pack, matching the
shape every descriptor took in 045.  as_vec<u8, len_from_field<"n">> reads
exactly as before; what changed is that the slot is no longer positional,
which is what lets a constraint and a bound join it later without inventing a
second spelling.

The migration the issue promised is empty, and that is the result rather than
a skipped step: every old spelling was already a valid one-entry pack, so no
schema in test/, examples/ or docs/ needed an edit.  The diff outside
include/, docs/ and single_header/ is three files, all additions.

Only the four tags that have a size change here.  The sizeless four keep
their arity until 050: giving them a pack now while rejecting the only entry
it could hold would specify a pack that admits nothing, and there is no
honest concept for that.

The pack machinery moves to field/field_options.hpp because type_tags.hpp
cannot include field_descriptors.hpp -- that closes a five-header loop
through match_case.  The first attempt pulled field_size_deduce.hpp along
with it and hit the same class of failure from a different edge, surfacing as
"struct_field_list_impl was not declared" three headers away, exactly the
confusion the guards were always going to produce.  The compile-time-only
deduction header is the one that has no field_list dependency.

as_trivial's requires-clause now reads the resolved pack size, keeping both
conjuncts: the fixed_size_like guard is what makes the fit check a
well-formed expression when the size is variable.  It still constrains --
rejects_oversized_as_trivial fails on field_fits_to_underlying_type itself,
not on something incidental, checked against a no-CASE control.

as_trivial<u32> now compiles, defaulting to sizeof, as basic_field already
did.
